middleware change, new services added.

+++ Application/Dtos/Rating/RatingDto.cs
@@ -11,10 +11,10 @@ namespace Application.Dtos.Rating
     public class RatingDto
     {
         public int? Id { get; set; }
-        public int? OverallRating { get; set; }
-        public int? PriceRating { get; set; }
-        public int? FemininityRating { get; set; }
-        public int? MasculinityRating { get; set; }
+        public float? OverallRating { get; set; }
+        public float? PriceRating { get; set; }
+        public float? FemininityRating { get; set; }
+        public float? MasculinityRating { get; set; }
         public UserDto? User { get; set; }  // Nested user details
         public FragranceDto? Fragrance { get; set; }  // Nested fragrance details
     }
